refactor(feedback): refactor ad-hoc
