+++ frontend/src/locales/fa.json
@@ -203,8 +203,8 @@
     "provideFeedback": "ثبت بازخورد",
     "writeFeedback": "نوشتن بازخورد",
     "writeFeedbackContent": "متن بازخورد را بنویسید.",
-    "evidence": "شواهد",
-    "writeEvidenceContent": "شواهد خود را وارد کنید.",
+    "evidence": "مصداق",
+    "writeEvidenceContent": "مصداق خود را وارد کنید.",
     "noOwnedRequests": "درخواستی برای نمایش وجود ندارد.",
     "noInvitedRequests": "درخواستی برای نمایش وجود ندارد.",
     "invitedRequests": "درخواست‌های دریافتی",
