+++ frontend/src/pages/index/adhoc-feedback.vue
@@ -1,16 +1,13 @@
 <script setup lang="ts">
 import { PHeading, PIconButton } from '@pey/core';
 import { PeyPlusIcon } from '@pey/icons';
-import { useRouteQuery } from '@vueuse/router';
 
-useHead({ title: 'درخواست فیدبک' });
+useHead({ title: 'فیدبک مستقیم' });
 
 const { t } = useI18n();
 const route = useRoute();
 
-const showHeader = computed(() => route.name === 'feedback');
-
-const currentTab = useRouteQuery<'owned' | 'invited'>('tab', 'owned');
+const showHeader = computed(() => route.name === 'adhoc-feedback');
 </script>
 
 <template>
@@ -25,12 +22,12 @@ const currentTab = useRouteQuery<'owned' | 'invited'>('tab', 'owned');
         <i class="i-mdi-account-supervisor text-h1 text-primary" />
 
         <PHeading level="h1" responsive>
-          {{ t('common.feedbackRequest') }}
+          {{ t('feedback.adhocFeedback') }}
         </PHeading>
       </div>
 
-      <div v-if="currentTab !== 'invited'" class="flex items-center gap-2">
-        <NuxtLink :to="{ name: 'feedback-new' }">
+      <div class="flex items-center gap-2">
+        <NuxtLink :to="{ name: 'adhoc-feedback-new' }">
           <PIconButton class="shrink-0" :icon="PeyPlusIcon" tabindex="-1" />
         </NuxtLink>
       </div>
