+++ frontend/src/pages/index/adhoc-feedback/new.vue
@@ -1,14 +1,14 @@
 <script setup lang="ts">
-import { PText } from '@pey/core';
+import { PBox, PText } from '@pey/core';
 import { useRouter } from 'vue-router';
 
-definePageMeta({ name: 'feedback-adhoc-new' });
+definePageMeta({ name: 'adhoc-feedback-new' });
 
 const { t } = useI18n();
 const router = useRouter();
 
 const handleSuccess = () => {
-  router.push({ name: 'feedback' });
+  router.push({ name: 'adhoc-feedback' });
 };
 
 const handleCancel = () => {
@@ -17,7 +17,7 @@ const handleCancel = () => {
 </script>
 
 <template>
-  <div>
+  <PBox class="mx-auto max-w-3xl bg-white px-2 py-8 sm:px-4 lg:px-8 lg:pt-10">
     <div class="mb-4 flex items-center gap-4 border-b border-gray-10 pb-4">
       <i class="i-mdi-comment-quote-outline text-h1 text-primary" />
       <PText as="h2" variant="h3" weight="bold">
@@ -25,5 +25,5 @@ const handleCancel = () => {
       </PText>
     </div>
     <FeedbackAdhocForm @success="handleSuccess" @cancel="handleCancel" />
-  </div>
+  </PBox>
 </template>
